Increase backup robustness
Since around a week the spotify API does return sporadic 401 errors for
some calls that succeed on retry. To compensate for this, 401 is added
to the list of status codes to retry.

Wrap the complete backup process into try/except blocks to make sure a
single failing backup function does not stop the whole process.